PlaceQueryService 생성하여 장소 검색 메소드 이동 및 PlanService에서 제거
